% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_value.R
\name{compute_correlation}
\alias{compute_correlation}
\title{Compute correlation between index value and single trait}
\usage{
compute_correlation(ptbl_bv, ptbl_index, ps_breed, ps_trait)
}
\arguments{
\item{ptbl_bv}{tibble with breeding values}

\item{ptbl_index}{tibble with index}

\item{ps_breed}{breed}

\item{ps_trait}{trait}
}
\value{
correlation between index and breeding values of single trait
}
\description{
Given a tibble of index values and a tibble of breeding values,
a trait name and a breed name, we select all breeding values for
the given trait and the given breed and compute the correlation
to the indices of the same animals
}
