% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_value.R
\name{compute_economic_value}
\alias{compute_economic_value}
\title{Compute Economic Values Based On Price Changes}
\usage{
compute_economic_value(pn_mean, pn_sd, pvec_class_freq = NULL,
  pvec_threshold = NULL, pvec_price, pn_delta_mean, pb_verbose = FALSE)
}
\arguments{
\item{pn_mean}{empirical population mean}

\item{pn_sd}{empirical population standard deviation}

\item{pvec_class_freq}{vector of empirical trait frequencies}

\item{pvec_threshold}{vector of thresholds given by payment system}

\item{pvec_price}{vector of prices given by payment system}

\item{pn_delta_mean}{small change of population mean}
}
\value{
ev_result result of economic value per pn_delta_mean
}
\description{
In this function, the economic value is computed as the change
in expected average price over a population for a given change
in the population mean. The computation
of the economic value is simplified by the assumption that the
costs can be kept constant for the different levels of the population
mean. Futhermore, we assume that the change in population mean
does not change the amount of product produced. The result is
standardized and expressed per unit of the trait.
}
