% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_value.R
\name{compute_ev_price_cont}
\alias{compute_ev_price_cont}
\title{Compute numeric value of economic value based on revenue}
\usage{
compute_ev_price_cont(pn_mean, pn_sd, pvec_threshold, pvec_price,
  pn_delta_mean, pn_gen_sd, pb_verbose = FALSE)
}
\arguments{
\item{pn_mean}{empirical population mean}

\item{pn_sd}{empirical population standard deviation}

\item{pvec_threshold}{vector of thresholds given by payment system}

\item{pvec_price}{vector of prices given by payment system}

\item{pn_delta_mean}{small change of population mean}

\item{pn_gen_sd}{genetic standard deviation of trait}

\item{pb_verbose}{level of output verbosity}
}
\value{
ev_result      list of two economic values, per unit trait and per genetic sd, if pn_gen_sd is given
}
\description{
This function computes the economic value based on changes in revenue for
a small change in the population mean.
}
