% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_value.R
\name{compute_ev_price_disc}
\alias{compute_ev_price_disc}
\title{Compute economic value based on price change for discrete trait}
\usage{
compute_ev_price_disc(pn_mean, pn_sd, pvec_class_freq, pvec_threshold,
  pvec_price, pn_delta_mean, pb_verbose = FALSE)
}
\arguments{
\item{pn_mean}{empirical population mean}

\item{pn_sd}{empirical population standard deviation}

\item{pvec_class_freq}{vector of empirical trait frequencies}

\item{pvec_threshold}{vector of thresholds given by payment system}

\item{pvec_price}{vector of prices given by payment system}

\item{pn_delta_mean}{small change of population mean}

\item{pb_verbose}{verbosity flag}
}
\value{
ev_result
}
\description{
Economic value is computed based on price change between base distribution
and a shifted distribution for a discrete trait.
}
