% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_value.R
\name{get_relative_economic_factors}
\alias{get_relative_economic_factors}
\title{Relative Economic Factors From Economic Values}
\usage{
get_relative_economic_factors(ptbl_economic_value,
  pb_first_row_trait_name = FALSE)
}
\arguments{
\item{ptbl_economic_value}{tibble with economic values}

\item{pb_first_row_trait_name}{flag indicating wether first column are trait names}
}
\value{
tbl_rel_fact tibble with relative factors
}
\description{
Given a tibble with economic values per genetic standard deviation in
parameter ptbl_economic_value, the relative weight of each economic
value with respect to the total sum of all economic values is computed.
}
