% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_value.R
\name{weight_economic_value}
\alias{weight_economic_value}
\title{Multiply Economic Values With A Weight Matrix}
\usage{
weight_economic_value(ptbl_economic_value, ptbl_weight,
  pb_first_row_trait_name = FALSE)
}
\arguments{
\item{ptbl_economic_value}{tibble with economic values}

\item{ptbl_weight}{tibble containing weights}

\item{pb_first_row_trait_name}{flag indicating wether first column are trait names}
}
\value{
tbl_weighted_result     weighted economic factors
}
\description{
Compute the weighted factors based on an input tibble with unweighted
economic values and a tibble with weighting factors. The two input
tibbles must have the same dimensions. If the first column contains
some names this is indicated with the parameter pb_first_row_trait_name
and the first column is ignored for the computation of the weighted
factors. In the result the first column is added back in.
}
